+++ src/generators/crud/Generator.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\gii\generators\crud;
@@ -24,6 +24,7 @@ use yii\web\Controller;
  * @property-read string $nameAttribute
  * @property-read string[] $searchAttributes Searchable attributes.
  * @property-read \yii\db\TableSchema|false $tableSchema
+ * @property-read string $viewPath The controller view path.
  *
  * @author Qiang Xue <qiang.xue@gmail.com>
  * @since 2.0