+++ src/generators/model/Generator.php
@@ -1,8 +1,8 @@
 <?php
 /**
- * @link http://www.yiiframework.com/
+ * @link https://www.yiiframework.com/
  * @copyright Copyright (c) 2008 Yii Software LLC
- * @license http://www.yiiframework.com/license/
+ * @license https://www.yiiframework.com/license/
  */
 
 namespace yii\gii\generators\model;
@@ -76,6 +76,7 @@ class Generator extends \yii\gii\Generator
      * @var string[]
      */
     protected $classNames = [];
+
 
     public function init()
     {