+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 gii extension Change Log
 2.2.5 under development
 -----------------------
 
+- Bug #500: Fix missing namespace in CRUD index template (mohamed-nazim)
 - Ehn #489: Added support for the `::class` constant in model generator via the `useClassConstant` setting (rhertogh)
 - Bug #502: Setting `skipOnEmpty` to fix "trim(): Passing null to parameter" in `generators/model/Generator.php` (rhertogh)